requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
kafka-python
python-decouple
requests